+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Latest
+
+- Add `autoprint/1` to automatically open the print dialog in a browser
+
 ## 0.6.1 (2023-01-19)
 
 - Fix bug with zero width strings and empty rows (also fixes [#24])